Move ImageAnnotationModalSVG to components folder
