+++ app/packs/src/components/ImageAnnotationModalSVG.js
